Update akaze.cpp
Update kaze.cpp

Revert "Update kaze.cpp"

This reverts commit e13f556560e27d555f6b9e736ca2ec136025d6bb.

Revert "Update akaze.cpp"

This reverts commit d162a2f8a9ebab395f15da40920672883b311541.

a

+++ modules/features2d/src/akaze.cpp
@@ -86,8 +86,8 @@ namespace cv
         void setDescriptorChannels(int dch) CV_OVERRIDE { descriptor_channels = dch; }
         int getDescriptorChannels() const CV_OVERRIDE { return descriptor_channels; }
 
-        void setThreshold(float threshold_) CV_OVERRIDE { threshold = threshold_; }
-        float getThreshold() const CV_OVERRIDE { return threshold; }
+        void setThreshold(double threshold_) CV_OVERRIDE { threshold = threshold_; }
+        double getThreshold() const CV_OVERRIDE { return threshold; }
 
         void setNOctaves(int octaves_) CV_OVERRIDE { octaves = octaves_; }
         int getNOctaves() const CV_OVERRIDE { return octaves; }
